requirements.txt
blinker==1.6.2
click==8.1.4
Flask==2.3.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-dotenv==1.0.0
Werkzeug==2.3.6
